+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - add: `StkTableColumn['type']` `expand`
   - add: slot `expand`
   - add: `StkTableColumn.customCell` new prop `expanded`
+  - add: `props.expandConfig`
 * change: remove deprecated type `CustomHeaderCellFunc` & `CustomCellFunc`
 
 ## 0.4.15